Fixed error on travis.yml
/usr/bin/g++   -DPACKAGE_VERSION=\"0.9.0\" -g -I/home/travis/build/xboxdrv/xboxdrv/src -I/home/travis/build/xboxdrv/xboxdrv/build-Debug/src -isystem /usr/include/dbus-1.0 -isystem /usr/lib/x86_64-linux-gnu/dbus-1.0/include -isystem /usr/include/glib-2.0 -isystem /usr/lib/x86_64-linux-gnu/glib-2.0/include -isystem /usr/include/libusb-1.0 -isystem /home/travis/build/xboxdrv/xboxdrv/external/googletest/googletest/include    -std=c++11 -pedantic -Wall -Wextra -Wno-c++0x-compat -Wnon-virtual-dtor -Weffc++ -Wconversion -Wshadow -Wcast-qual -Winit-self -Wno-unused-parameter -Werror -o CMakeFiles/libxboxdrv.dir/src/ui_event_emitter.cpp.o -c /home/travis/build/xboxdrv/xboxdrv/src/ui_event_emitter.cpp
/home/travis/build/xboxdrv/xboxdrv/src/xpad_device.cpp:150:52: error: conversion to ‘int’ from ‘long unsigned int’ may alter its value [-Werror=conversion]
 const int xpad_devices_count = sizeof(xpad_devices)/sizeof(XPadDevice);

+++ src/xpad_device.cpp
@@ -147,7 +147,7 @@ XPadDevice xpad_devices[] = {
   { GAMEPAD_HAMA_CRUX,        0x1038, 0x0310, "Hama cruX Gaming Keyboard" }
 };
 
-const int xpad_devices_count = sizeof(xpad_devices)/sizeof(XPadDevice);
+const int xpad_devices_count = static_cast<int>(sizeof(xpad_devices)/sizeof(XPadDevice));
 
 bool find_xpad_device(uint16_t idVendor, uint16_t idProduct, XPadDevice* dev_type)
 {
